backend/requirements.txt
django
djangorestframework
pymysql
django-cors-headers
pyjwt
pillow
requests
